Move notification check logic to ConnectivityToolViewModel+Notifications
Extract testNotifications, Jetpack/permission/config sub-checks,
enableOrderNotifications, error enums, and related localization
into a dedicated extension file to reduce view model size.
